Make the 'share' format option emit shared memory resident size
The 'share' option has been unused for the lifetime of procps-ng.

Co-opt it to print the resident shared memory size of a process,
/proc/$pid/status's RssShare field (kB). That's equivalent to
/proc/$pid/statm's share field (pages).

This allows system operators to determine the proportion of a process's
memory that is potentially shared with other processes.

The next patch will build on this by exposing the private (unshared)
resident memory of a process as a separately printable and sortable
field.

The presence of memory in the 'shared' column does not necessarily mean
the memory is actually shared, since the process may allocate shared
memory with nothing attached to it. It is only share-able.

Note that Linux has facilities to expose proportional allocation of
shared memory costs between processes, as can be seen in the Pss_Anon,
Pss_File and Pss_Shmem fields of /proc/$pid/smaps_rollup, but this is
not presently exposed by 'ps' and can be left for a later step.